Move DiskIndexWithDestructorCleanup to separate files.
